Move the last-DataNode removal test to its own 1C1D suite class
failWhenRemovingLastSingleReplicaDataNodeUseSQL only needs a 1C1D
cluster, but it lived in IoTDBRemoveDataNodeNormalIT which is
@Category({ClusterIT.class}) and therefore ran in the 1C3D suite
alongside the multi-DataNode removal tests.

Extract it into a new IoTDBRemoveLastDataNodeIT annotated
@Category({LocalStandaloneIT.class}) so it runs in the 1C1D (Simple)
suite, isolated from the cluster tests. Verified with the default
with-integration-tests profile (no -PClusterIT): Tests run: 1,
Failures: 0 on a 1C1D cluster.
