+++ integration-test/src/test/java/org/apache/iotdb/confignode/it/removedatanode/IoTDBRemoveDataNodeNormalIT.java
@@ -136,47 +136,6 @@ public class IoTDBRemoveDataNodeNormalIT {
   //  }
 
   @Test
-  public void failWhenRemovingLastSingleReplicaDataNodeUseSQL() throws Exception {
-    // With a single replica (schema_replication_factor and data_replication_factor are both 1),
-    // removing DataNodes is still supported as long as more than one DataNode remains, but the last
-    // remaining DataNode cannot be removed because there is nowhere to migrate its regions to.
-    // Here we set up 1C1D with single replica and try to remove the only DataNode, which must fail
-    // because removing it would leave the cluster with no DataNode.
-    EnvFactory.getEnv()
-        .getConfig()
-        .getCommonConfig()
-        .setDataRegionConsensusProtocolClass(ConsensusFactory.IOT_CONSENSUS)
-        .setSchemaReplicationFactor(1)
-        .setDataReplicationFactor(1)
-        .setDefaultDataRegionGroupNumPerDatabase(1);
-    EnvFactory.getEnv().initClusterEnvironment(1, 1);
-
-    try (final Connection connection = makeItCloseQuietly(EnvFactory.getEnv().getConnection());
-        final Statement statement = makeItCloseQuietly(connection.createStatement());
-        final ResultSet resultSet = statement.executeQuery(SHOW_DATANODES)) {
-      final Set<Integer> allDataNodeId = new HashSet<>();
-      while (resultSet.next()) {
-        allDataNodeId.add(resultSet.getInt(ColumnHeaderConstant.NODE_ID));
-      }
-
-      final String removeDataNodeSQL =
-          generateRemoveString(selectRemoveDataNodes(allDataNodeId, 1));
-      try {
-        statement.execute(removeDataNodeSQL);
-        Assert.fail(
-            "Remove DataNode should fail when it would leave no DataNode under single replica");
-      } catch (final IoTDBSQLException e) {
-        // The unified rejection message reports the gap and, for a single replica, appends the
-        // "at least one DataNode must always remain" hint.
-        Assert.assertTrue(e.getMessage(), e.getMessage().contains("Cannot remove"));
-        Assert.assertTrue(e.getMessage(), e.getMessage().contains("single replica"));
-        Assert.assertFalse(
-            e.getMessage(), e.getMessage().contains("Failed to remove all requested data nodes"));
-      }
-    }
-  }
-
-  @Test
   public void fail1C3DTestIoTUseSQL() throws Exception {
     // Setup 1C3D with schema replication factor = 3, and remove 1D, this test should fail due to
     // insufficient DN for holding schema
